+++ cloudflare/Get-NewDnsEntries.ps1
@@ -1,4 +1,4 @@
-﻿$postgresService = $( kubectl get service "postgres-service" -n "event-triangle" -o json ) | ConvertFrom-Json
+$postgresService = $( kubectl get service "postgres-service" -n "event-triangle" -o json ) | ConvertFrom-Json
 $postgresPublicIp = $postgresService.status.loadBalancer.ingress[0].ip
 
 $rabbitService = $( kubectl get service "event-rabbitmq" -n "event-triangle" -o json ) | ConvertFrom-Json